requirements.txt
streamlit
networkx
pyvis
pandas
pdfplumber
python-dotenv
langchain-text-splitters
instructor
scikit-learn
numpy
google-genai
jsonref
langchain
langchain-core
langchain-community
rapidfuzz
nltk
sentence-transformers
fuzzywuzzy
python-Levenshtein
ipython
torch
pydantic
